fix(kernel): stop a single float64 fact from killing the whole fixpoint (F-NUM-1)
A live dogfood session logged this 1257 times, ~4 per 2s, for its entire run:

    evaluate: fixpoint evaluation failed: value 110 (4) is not a number

Type tag 4 is ast.Float64Type. Three things compound into a silent outage:

1. types.Fact.ToAtom is Decl-blind: every Go float64 becomes ast.Float64,
   whatever the predicate declares.
2. The pinned Mangle fork implements <, <=, >, >= over int64 ONLY. builtin.go
   routes each through getNumberValues -> getNumberValue, which errors on any
   constant whose Type != NumberType. getFloatValue exists but has no caller —
   there is no float comparison in the language at all.
3. That error propagates out of EvalStratifiedProgram, so evaluate() returns at
   kernel_eval.go:299 BEFORE assigning k.store. One bad fact means the entire
   kernel derives nothing, on every subsequent pass — and the message names only
   the value, never the predicate carrying it.

The corpus declares zero /float64 bounds, so a Float64 in the EDB is always a bug
here. On a system whose premise is that logic is the executive, this took the
executive offline while the TUI carried on as if nothing were wrong.

Boundary fix (the general one):
  factToAtomLocked applies Decl-directed coercion at every path that fills
  cachedAtoms — addFactIfNewLocked, the evaluateFullLocked rebuild, the
  in-place update path, and factsToAtomsLocked (whose stated invariant is to be
  bit-identical to the full path, so it had to coerce too). Integral float in a
  /number slot narrows to ast.Number; non-integral is rejected naming the
  predicate and arg index. evaluateFullLocked now evicts an unconvertible fact
  and keeps facts/cachedAtoms/factIndex in lockstep rather than aborting.
  New atomCacheStale flag forces one reconversion after a policy rebuild and for
  facts admitted before programInfo exists (no Decls yet at boot).

Source fixes:
  - normalizePercent returned float64 into tool_learning (/number).
  - feedback.go wrote %.2f into GENERATED MANGLE SOURCE persisted to learned.mg,
    so a float poisoned every later boot too.
  - decomposer.go fed raw 0..1 ratios into seven intelligence_* /number slots.
  All now route through the new types.PercentScale.

Corpus fix:
  17 rules compared /number args against float literals (AvgQuality < 50.0,
  Confidence > 0.7, Coverage < 0.3, Weight > 0.5, ...) in autopoiesis.mg and
  intelligence.mg. Each was a latent whole-kernel outage, armed the moment a
  matching fact appeared. Rescaled to integer percent.

Tests:
  - TestFactBoundary_* — red proof reproduces the exact production error
    ("value 10 (4) is not a number") and shows one bad fact killing an unrelated
    context_priority query.
  - TestCorpus_NoFloatLiteralInComparison — static ban; the regex is itself
    tested for both operand positions and against false positives.
  - TestPercentScale* — includes a guard that a bare float64 still maps to
    Float64Type, so the helper's rationale is revisited if ToAtom ever changes.

Residual, recorded in the ledger, not swept: virtual predicates built by
appendAtom bypass k.facts and so the coercion; strategic_knowledge is asserted
with no Decl anywhere; internal/tactile asserts test_coverage with 2 args against
an arity-1 Decl.

+++ internal/autopoiesis/utils_coverage_test.go
@@ -11,17 +11,20 @@ func compilePatternForTest(pattern string) *regexp.Regexp {
 
 // --- normalizePercent ---
 
+// want is int64, not float64: tool_learning's numeric slots are declared
+// /number, and this Mangle fork compares int64 only — a float here aborts the
+// whole kernel fixpoint rather than just failing the tool_quality_* rules.
 func TestNormalizePercent_AllBands(t *testing.T) {
 	tests := []struct {
 		name  string
 		input float64
-		want  float64
+		want  int64
 	}{
 		{"zero", 0.0, 0},
 		{"negative", -1.0, 0},
 		{"fraction", 0.5, 50},
 		{"fraction_low", 0.1, 10},
-		{"one", 1.0, 1.0},   // >= 1, but not > 100
+		{"one", 1.0, 1},     // >= 1, read as a percent, not a ratio
 		{"fifty", 50.0, 50}, // >= 1, <= 100
 		{"hundred", 100.0, 100},
 		{"over_hundred", 200.0, 100},
